Move common test utils to common directory
Move testutils and checker to the common directory to make their functions clear and promote reuse. So we have test files structured in a way resembling the source

[ghstack-poisoned]
