+++ onnxscript/test/common/__init__.py
@@ -1,0 +1,1 @@
+"""Share components for testing."""
