+++ onnxscript/test/common/testutils.py
@@ -8,7 +8,7 @@ import unittest
 from onnx import FunctionProto, GraphProto, ModelProto, parser
 
 from onnxscript import OnnxFunction
-from onnxscript.test.checker import isomorphic
+from onnxscript.test.common.checker import isomorphic
 
 
 def function_proto(f):
