+++ onnxscript/test/operator_test.py
@@ -10,7 +10,7 @@ from onnx import helper
 from onnxscript import script
 from onnxscript.onnx_opset import opset15 as op
 from onnxscript.onnx_types import FLOAT
-from onnxscript.test.testutils import TestBase
+from onnxscript.test.common.testutils import TestBase
 
 
 class TestConverter(TestBase):
