+++ onnxscript/test/type_annotation_test.py
@@ -8,7 +8,7 @@ import unittest
 from onnxscript import script
 from onnxscript.onnx_opset import opset15 as op
 from onnxscript.onnx_types import FLOAT
-from onnxscript.test.testutils import TestBase
+from onnxscript.test.common.testutils import TestBase
 
 
 class TypeAnnotationTester(TestBase):
